% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoHAB.R
\name{EcoHAB}
\alias{EcoHAB}
\title{R6 class processing Eco-HAB experimental data}
\description{
This class stores raw data and convert them into events and summaries.
}
\details{
The EcoHAB class is generated from a list of raw data files, a ID file and a time line file, and store them as private members. External access is granted by active bindings to protect the reference nature of R6 class and data.table. Analysis results are stored as private and accessed by functions that return a copy of data.table.
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{raw}}{Processed raw data in a data.table (read-only)}

    \item{\code{idlist}}{Animal ID information in a data.table (read-only)}

    \item{\code{timeline}}{Experimental time line setting in a data.table (read-only)}

    \item{\code{timeline_bin}}{Subdivided time line setting in a data.table (read-only)}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-EcoHAB-initialize}{\code{EcoHAB$new()}}
    \item \href{#method-EcoHAB-dummy}{\code{EcoHAB$dummy()}}
    \item \href{#method-EcoHAB-set_binsize}{\code{EcoHAB$set_binsize()}}
    \item \href{#method-EcoHAB-calc_events}{\code{EcoHAB$calc_events()}}
    \item \href{#method-EcoHAB-edit_events}{\code{EcoHAB$edit_events()}}
    \item \href{#method-EcoHAB-calc_single}{\code{EcoHAB$calc_single()}}
    \item \href{#method-EcoHAB-calc_pair}{\code{EcoHAB$calc_pair()}}
    \item \href{#method-EcoHAB-calc_follow}{\code{EcoHAB$calc_follow()}}
    \item \href{#method-EcoHAB-calc_activity}{\code{EcoHAB$calc_activity()}}
    \item \href{#method-EcoHAB-all_results}{\code{EcoHAB$all_results()}}
    \item \href{#method-EcoHAB-get_result}{\code{EcoHAB$get_result()}}
    \item \href{#method-EcoHAB-reset_results}{\code{EcoHAB$reset_results()}}
    \item \href{#method-EcoHAB-plot_occupancy}{\code{EcoHAB$plot_occupancy()}}
    \item \href{#method-EcoHAB-clone}{\code{EcoHAB$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-initialize}{}}}
\subsection{\code{EcoHAB$new()}}{
  Create a new Eco-HAB data set from input files
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$new(
  raw_files = NULL,
  idfile = NULL,
  timefile = NULL,
  timezone = Sys.timezone()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{raw_files}}{A list of RFID recording text files}
      \item{\code{idfile}}{A text file containing animal ID information}
      \item{\code{timefile}}{A text file containing experimental time line information}
      \item{\code{timezone}}{The time zone of RFID recording}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-dummy"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-dummy}{}}}
\subsection{\code{EcoHAB$dummy()}}{
  Create a dummy Eco-HAB data set from pre-set parameters
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$dummy(
  n_subject = 10,
  p_step = 0.5,
  mean_step = 1000,
  sd_step = 100,
  t = 86400,
  p_cage = 0.99,
  alpha_cage = 10,
  alpha_tube = 2,
  theta = 1,
  timezone = "UTC",
  seed = 123
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{n_subject}}{Number of animals}
      \item{\code{p_step}}{Probability of animal moving forwards on the RFID reader chain}
      \item{\code{mean_step}}{Mean value of total movements}
      \item{\code{sd_step}}{Standard deviation of movements}
      \item{\code{t}}{Total recording time}
      \item{\code{p_cage}}{Probability of time stayed in cages}
      \item{\code{alpha_cage}}{Alpha of individual cage event time distribution (gamma)}
      \item{\code{alpha_tube}}{Alpha of individual tube event time distribution (gamma)}
      \item{\code{theta}}{Theta of individual cage or tube event time distribution (gamma)}
      \item{\code{timezone}}{The time zone of dummy RFID recording}
      \item{\code{seed}}{Random seed to generate dummy data}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The current EcoHAB object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-set_binsize"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-set_binsize}{}}}
\subsection{\code{EcoHAB$set_binsize()}}{
  Divide the experimental time line according to bin size
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$set_binsize(binsize)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{binsize}}{The duration of each bin in seconds}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The current EcoHAB object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-calc_events"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-calc_events}{}}}
\subsection{\code{EcoHAB$calc_events()}}{
  Populate events from RFID recordings
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$calc_events(threshold = 2)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{threshold}}{The maximal time the animal stays at a RFID reader in seconds}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The current EcoHAB object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-edit_events"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-edit_events}{}}}
\subsection{\code{EcoHAB$edit_events()}}{
  Show and edit events in a ShinyApp
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$edit_events()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-calc_single"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-calc_single}{}}}
\subsection{\code{EcoHAB$calc_single()}}{
  Populate the single events from individual events
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$calc_single(loc = c("cages", "all"))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{loc}}{The locations where single events are counted}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The current EcoHAB object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-calc_pair"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-calc_pair}{}}}
\subsection{\code{EcoHAB$calc_pair()}}{
  Populate the pair-wise events from individual events
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$calc_pair(loc = c("cages", "all"))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{loc}}{The locations where pair-wise events are counted}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The current EcoHAB object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-calc_follow"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-calc_follow}{}}}
\subsection{\code{EcoHAB$calc_follow()}}{
  Populate the following events from individual events
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$calc_follow(
  loc = c("tubes", "all"),
  mode = c("in_place", "delayed"),
  threshold = 2
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{loc}}{The locations where following events are counted}
      \item{\code{mode}}{The mode of following events identification}
      \item{\code{threshold}}{The maximal interval bewtween two animals in the "delayed" mode}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The current EcoHAB object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-calc_activity"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-calc_activity}{}}}
\subsection{\code{EcoHAB$calc_activity()}}{
  Calculate the counts and duration of given events in each time bin
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$calc_activity(name = c("events", "single", "pair", "follow"))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{name}}{The name of events data.table in the "results" list}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The current EcoHAB object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-all_results"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-all_results}{}}}
\subsection{\code{EcoHAB$all_results()}}{
  Get the names of all results
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$all_results()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A vector containing names of all elements in the "results" list
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-get_result"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-get_result}{}}}
\subsection{\code{EcoHAB$get_result()}}{
  Get a result in data.table
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$get_result(result_name)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{result_name}}{The name of events or analyses data.table in the "results" list}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A copy of data.table in the "results" list
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-reset_results"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-reset_results}{}}}
\subsection{\code{EcoHAB$reset_results()}}{
  Delete all analysis results
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$reset_results(prompt = "Delete all results?: (y/n): ")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{prompt}}{Instructions shown in the console}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    TRUE if results are deleted or FALSE otherwise
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-plot_occupancy"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-plot_occupancy}{}}}
\subsection{\code{EcoHAB$plot_occupancy()}}{
  Plot the cage occupancy in heatmap
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$plot_occupancy()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A "ggplot2" object containing the plot
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EcoHAB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EcoHAB-clone}{}}}
\subsection{\code{EcoHAB$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{EcoHAB$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
